refactor: not-found page move UI into page
