+++ app/not-found.tsx
@@ -1,7 +1,20 @@
 'use client';
 
-import NotFoundPage from '@/src/pages/404/NotFound';
+import Button from '@/src/components/UI/button/Button';
+import styles from '@/app/not-found/not-found.module.css';
 
 export default function Page(): React.ReactNode {
-  return <NotFoundPage />;
+  const handleClick = (): void => {
+    window.history.back();
+  };
+
+  return (
+    <div className={styles.not_found}>
+      <h2 className={styles.title}>404 - Not Found</h2>
+      <p className={styles.subtitle}>
+        This is not the page you&apos;re looking for...
+      </p>
+      <Button onClick={handleClick}>Back to base</Button>
+    </div>
+  );
 }
